Trait xtensa_lx::mutex::Mutex

pub trait Mutex {
    type Data;

    // Required method
    fn lock<R>(&mut self, f: impl FnOnce(&mut Self::Data) -> R) -> R;
}

Any object implementing this trait guarantees exclusive access to the data contained within the mutex for the duration of the lock.

Required Associated Types

type Data

Data protected by the mutex.

Required Methods

fn lock<R>(&mut self, f: impl FnOnce(&mut Self::Data) -> R) -> R

Creates a critical section and grants temporary access to the protected data.

Object Safety

This trait is not object safe.